Django/pjts/pjt03/requirements.txt
asgiref==3.5.2
Django==3.2.11
django-extensions==3.2.1
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
